# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "taskcube",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "v1.0"
    },
    "host": "localhost:8080",
    "basePath": "/taskcube/api/v1",
    "paths": {
...
        "/template": {
            "get": {
                "description": "查询任务模版",
                "tags": [
                    "任务模版"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/services.TaskTemplateService"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新任务模版",
                "tags": [
                    "任务模版"
                ],
                "parameters": [
                    {
                        "description": "TaskTemplateService",
                        "name": "TaskTemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TaskTemplateService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.TaskTemplateService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建任务模版",
                "tags": [
                    "任务模版"
                ],
                "parameters": [
                    {
                        "description": "TaskTemplateService",
                        "name": "TaskTemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TaskTemplateService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.TaskTemplateService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/template/{template}": {
            "delete": {
                "description": "删除任务模版",
                "tags": [
                    "任务模版"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "template",
                        "name": "template",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/template/{template}/submit": {
            "post": {
                "description": "提交任务模版",
                "tags": [
                    "任务模版"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "template",
                        "name": "template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "返回码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string"
                },
                "message": {
                    "description": "返回信息",
                    "type": "string"
                }
            }
        },
...
        "services.TaskTemplateService": {
            "type": "object",
            "properties": {
                "kind": {
                    "$ref": "#/definitions/types.KindType"
                },
                "metadata": {
                    "$ref": "#/definitions/types.TaskMetadata"
                },
                "spec": {
                    "$ref": "#/definitions/types.TaskTemplateSpec"
                }
            }
        },
        "types.KindType": {
            "type": "string",
            "enum": [
                "Script",
                "Task",
                "TaskTemplate"
            ],
            "x-enum-varnames": [
                "ScriptKind",
                "TaskKind",
                "TaskTemplateKind"
            ]
        },
...
        "types.TaskMetadata": {
            "type": "object",
            "properties": {
                "createAt": {
                    "type": "string"
                },
                "deleteAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "types.TaskSpecStep": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "input": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "script": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "types.TaskTemplateSpec": {
            "type": "object",
            "properties": {
                "input": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.TaskSpecStep"
                    }
                }
            }
        }
    }
}